Fix keep-alive disconnects on system time changes
mosquitto_time_init must be called before calling mosquitto_time,
otherwise time_clock will be 0 in mosquitto_time which will cause the
realtime clock to be used when the intention seems to be to use the
best monotonic clock supported by the platform.

That in turn causes the broker to kick clients when system time is
changed because it's misinterpreted as clients failing on their
keep-alive promises.