Fix CMake builds with WITH_TLS=OFF
Recent CMake changes caused CMake builds with the WITH_TLS option set to OFF to fail. The OpenSSL package is only found (find_package()) if WITH_TLS is ON, but linking to OpenSSL for the broker and library is not guarded by WITH_TLS. The build therefore fails.

Guard linking to OpenSSL, only linking if WITH_TLS is set.