Fix expiry list corruption
The internal expiry list can be corrupted if a client connects with a
nonzero session expiration time, they disconnect, a plugin kicks said
client, and mosquitto expires their session (in that order). The issue
arises because the client context is added to the expiry_list twice. Fix
this by first checking if the client is already on the list before
adding them again.

Signed-off-by: Nicholas Jackson <nickajacks1@gmail.com>