Adding new means for brigde authentication.
This patch introduces a new way to provide credentials for bridge
authentication.

Currently the only way to provide bridge authentication credentials
is to set remote_username and remote_password in the configuration
file. There was no other way to provide these credentials.

But for example on some platforms these credentials are stored in
a TPM. There is no direct way to pull these credentials from TPM and
use it for bridge authentication.

With this patch these credentials can be provided by a plugin that
yields username/password on bridge setup and will be used for
broker authentication. Customers can provide this plugin
without publishing their source code.

For more details see: doc/brigde-authentication-by-plugin.md

Signed-off-by: Thorsten Wendt <thorsten.wendt@ebblo.com>